// mobile/src/shared/types.ts
export interface UserProfileDefaults {
  currency: string;
  location: string;
  companyType: string;
  industry: string;
  authorTitle: string;
  citRate: number;
  rssbRate: number;
  healthInsuranceRate: number;
  maternityRate: number;
  discountRate: number;
  loanRate: number;
  loanTermYears: number;
}